third_party/blink/web_tests/platform/generic/fast/dom/Range/detach-no-op-expected.txt

Tests that Range.detach() is a no-op.

On success, you will see a series of "PASS" messages, followed by "TEST COMPLETE".

PASS range.startContainer is link
PASS range.startOffset is 0
PASS range.endContainer is link
PASS range.endOffset is 0
PASS range.collapsed is true
PASS range.commonAncestorContainer is link
PASS range.detach() did not throw exception.
PASS range.startContainer is link
PASS range.startOffset is 0
PASS range.endContainer is link
PASS range.endOffset is 0
PASS range.collapsed is true
PASS range.commonAncestorContainer is link
PASS successfullyParsed is true

TEST COMPLETE

testlink
